Add authentication module
