+++ lib/quaderno-ruby.rb
@@ -1,7 +1,7 @@
 require 'ostruct'
 
 require 'quaderno-ruby/exceptions/exceptions'
-%w(crud deliver payment retrieve).each { |filename| require "quaderno-ruby/behavior/#{ filename }" }
+%w(authentication crud deliver payment retrieve).each { |filename| require "quaderno-ruby/behavior/#{filename}" }
 %w(base contact item invoice receipt credit estimate expense recurring document_item evidence payment webhook tax).each { |filename| require "quaderno-ruby/#{ filename }" }
 
 module Quaderno
